fully separate endpoint resolution from endpoint use for report generation
